+++ src/Components/DropDown/DropDown.module.scss
@@ -1,0 +1,41 @@
+.header {
+  background: none;
+  border: none;
+  border-radius: 10%;
+  box-shadow: 0px 0px 20px rgba($color: #000000, $alpha: 0.1);
+  width: 200px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  cursor: pointer;
+  font-size: 1.5rem;
+  padding-left: 1rem;
+  margin: 1vw 1rem;
+  font-weight: 500;
+  color: var(--grey-blue-2);
+  .arrow {
+    transition: 250ms linear;
+  }
+}
+.options {
+  background-color: var(--light-blue);
+  width: 200px;
+  font-size: 1.2rem;
+  margin: 1vw 1rem;
+  .optbtn {
+    display: inline-block;
+    width: inherit;
+    cursor: pointer;
+    font-weight: 400;
+    padding: 1vw 0 1vw 1rem;
+    background: none;
+    border: none;
+    text-align: left;
+    font-size: 1.2rem;
+    color: var(--base-color);
+  }
+  .optbtn:hover {
+    background-color: var(--light-green);
+    border-radius: 5%;
+  }
+}
