+++ src/Components/DropDown/Options.jsx
@@ -1,0 +1,26 @@
+import style from "./DropDown.module.scss";
+
+const Options = ({ data, setIndex, setDrop }) => {
+  const handleClick = (index) => {
+    setIndex(index);
+    setDrop(false);
+  };
+
+  return (
+    <div className={style.options}>
+      {data.map((opt, index) => {
+        return (
+          <button
+            key={opt.key}
+            className={style.optbtn}
+            onClick={() => handleClick(index)}
+          >
+            {opt.title}
+          </button>
+        );
+      })}
+    </div>
+  );
+};
+
+export default Options;
