--- src/Components/TimetableContent/MobileResponsive/Dropdown.module.scss
@@ -1,72 +1,0 @@
-.container {
-  display: none;
-  .header {
-    background: none;
-    border: none;
-    border-radius: 10%;
-    box-shadow: 0px 0px 20px rgba($color: #000000, $alpha: 0.1);
-    width: 220px;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    cursor: pointer;
-    font-size: 1.5rem;
-    padding-left: 1rem;
-    margin: 1vw 1rem;
-    font-weight: 500;
-    color: var(--grey-blue-2);
-    .arrow {
-      transition: 250ms linear;
-    }
-  }
-}
-
-@media screen and (width < 940px) {
-  .container {
-    display: block;
-  }
-}
-.options {
-  background-color: var(--light-blue);
-  width: 220px;
-  font-size: 1.2rem;
-  margin: 1vw 1rem;
-  .optbtn {
-    display: inline-block;
-    width: inherit;
-    cursor: pointer;
-    font-weight: 400;
-    padding: 1vw 0 1vw 1rem;
-    background: none;
-    border: none;
-    text-align: left;
-    font-size: 1.2rem;
-    color: var(--base-color);
-  }
-  .optbtn:hover {
-    background-color: var(--light-green);
-    border-radius: 5%;
-  }
-}
-
-.gridContainer {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-}
-
-.card {
-  padding: 2vw 2vw 0;
-  margin: 2vw 2vw 0;
-  span {
-    font-size: 1.5rem;
-    font-weight: 500;
-    padding-left: 2vw;
-  }
-}
-
-@media screen and (width < 500px) {
-  .card {
-    padding: 0;
-    margin: 1vw 1vw 0;
-  }
-}
