+++ src/Components/TimetableContent/TimetableContent.jsx
@@ -8,9 +8,9 @@ const TimetableContent = ({ data, err }) => {
 
   return (
     <>
-      {data.map(({ day, dishes }) => {
+      {data.map(({ day, dishes, key }) => {
         return (
-          <tr className={style.content}>
+          <tr key={key} className={style.content}>
             <th>{day}</th>
             {dishes.map((dish) => {
               return (
