+++ src/Pages/TimeTable/TimeTable.jsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from "react";
-import { TimetableContent, Dropdown } from "../../Components";
+import { TimetableContent, TableMb } from "../../Components";
 import style from "./TimeTable.module.scss";
 
 const TimeTable = () => {
@@ -16,7 +16,7 @@ const TimeTable = () => {
   return (
     <div className={style.container}>
       <h2>Live TimeTable</h2>
-      <Dropdown data={data} />
+      <TableMb data={data} />
       <table>
         <tr>
           <th className={style.radius1}> </th>
